requirements.txt
# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Graph Neural Networks
torch-geometric>=2.3.0
torch-scatter>=2.1.0
torch-sparse>=0.6.0
torch-cluster>=1.6.0

# Data processing and visualization
networkx>=3.0
matplotlib>=3.6.0
seaborn>=0.12.0
scikit-learn>=1.2.0

# Code processing
transformers>=4.21.0

# Data processing (already used)
requests>=2.28.0
unidiff>=0.7.0
tqdm>=4.65.0

# Code quality and analysis (NEW for data curation)
radon>=6.0.1  # Code complexity metrics
pylint>=3.0.0  # Code quality analysis

# Experiment tracking and configuration
mlflow>=2.8.0
pyyaml>=6.0